Created sequence diagram
This diagram shows the flow of an assignment submission. A student submits the assignment to the web page (frontend) which is then sent to the backend for validation (checking size and type of file etc..). The information is then shared with the teacher, and the grade is updated on the front end for the student to see.